feat: view 로직 구성
