+++ src/main/kotlin/step2/strategy/calculable/Calculable.kt
@@ -1,4 +1,4 @@
-package step2.calculable
+package step2.strategy.calculable
 
 @FunctionalInterface
 interface Calculable {
