+++ src/main/kotlin/step2/strategy/input/ConsoleInputStrategy.kt
@@ -1,0 +1,5 @@
+package step2.strategy.input
+
+class ConsoleInputStrategy : InputStrategy {
+    override fun execute(): String? = readLine()
+}
