+++ src/main/kotlin/step2/strategy/output/ConsoleOutputStrategy.kt
@@ -1,0 +1,5 @@
+package step2.strategy.output
+
+class ConsoleOutputStrategy : OutputStrategy {
+    override fun execute(result: String) = println(result)
+}
