+++ src/main/kotlin/step2/strategy/output/OutputStrategy.kt
@@ -1,0 +1,6 @@
+package step2.strategy.output
+
+@FunctionalInterface
+interface OutputStrategy {
+    fun execute(result: String)
+}
