+++ src/test/kotlin/step2/strategy/calculable/strategy/MultiplyCalculableStrategyTest.kt
@@ -1,4 +1,4 @@
-package step2.calculable.strategy
+package step2.strategy.calculable.strategy
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Assertions.assertAll
@@ -7,27 +7,27 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
 
-class ModCalculableStrategyTest {
+class MultiplyCalculableStrategyTest {
 
-    @DisplayName("ModCalculableStrategy 인스턴스 생성 여부 테스트")
+    @DisplayName("MultiplyCalculableStrategy 인스턴스 생성 여부 테스트")
     @Test
     fun constructor_test() {
         // given and when
-        val calculable = ModCalculableStrategy()
+        val calculable = MultiplyCalculableStrategy()
 
         // then
         assertAll(
             { assertThat(calculable).isNotNull },
-            { assertThat(calculable).isExactlyInstanceOf(ModCalculableStrategy::class.java) },
+            { assertThat(calculable).isExactlyInstanceOf(MultiplyCalculableStrategy::class.java) },
         )
     }
 
-    @DisplayName("ModCalculableStrategy 인스턴스 calculate() 기능 테스트")
+    @DisplayName("MultiplyCalculableStrategy 인스턴스 calculate() 기능 테스트")
     @ParameterizedTest
-    @CsvSource(value = ["1:1:0", "-2:2:0", "-1:-1:0"], delimiter = ':')
+    @CsvSource(value = ["1:1:1", "-2:2:-4", "-1:-1:1"], delimiter = ':')
     fun calculate_test(x: Int, y: Int, expected: Int) {
         // given
-        val calculable = ModCalculableStrategy()
+        val calculable = MultiplyCalculableStrategy()
 
         // when
         val actual = calculable.calculate(x, y)
