+++ src/test/kotlin/step2/strategy/calculable/strategy/SubtractCalculableStrategyTest.kt
@@ -1,4 +1,4 @@
-package step2.calculable.strategy
+package step2.strategy.calculable.strategy
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Assertions.assertAll
@@ -7,27 +7,27 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
 
-class AddCalculableStrategyTest {
+class SubtractCalculableStrategyTest {
 
-    @DisplayName("AddCalculableStrategy 인스턴스 생성 여부 테스트")
+    @DisplayName("SubtractCalculableStrategy 인스턴스 생성 여부 테스트")
     @Test
     fun constructor_test() {
         // given and when
-        val calculable = AddCalculableStrategy()
+        val calculable = SubtractCalculableStrategy()
 
         // then
         assertAll(
             { assertThat(calculable).isNotNull },
-            { assertThat(calculable).isExactlyInstanceOf(AddCalculableStrategy::class.java) },
+            { assertThat(calculable).isExactlyInstanceOf(SubtractCalculableStrategy::class.java) },
         )
     }
 
-    @DisplayName("AddCalculableStrategy 인스턴스 calculate() 기능 테스트")
+    @DisplayName("SubtractCalculableStrategy 인스턴스 calculate() 기능 테스트")
     @ParameterizedTest
-    @CsvSource(value = ["1:1:2", "-2:2:0", "-1:-1:-2"], delimiter = ':')
+    @CsvSource(value = ["1:1:0", "-2:2:-4", "-1:-1:0"], delimiter = ':')
     fun calculate_test(x: Int, y: Int, expected: Int) {
         // given
-        val calculable = AddCalculableStrategy()
+        val calculable = SubtractCalculableStrategy()
 
         // when
         val actual = calculable.calculate(x, y)
